seperated AccountScreen into multiple components
