+++ components/ChangeEmail.js
@@ -1,0 +1,12 @@
+import React from "react";
+import { Button, Title, TextInput } from "react-native-paper";
+
+const ChangeEmail = () => {
+  return (
+    <>
+      <Title>Email</Title>
+    </>
+  );
+};
+
+export default ChangeEmail;
