+++ components/ChangePassword.js
@@ -1,0 +1,11 @@
+import React from "react";
+import { Button, Title, TextInput } from "react-native-paper";
+const ChangePassword = () => {
+  return (
+    <>
+      <Title>Password</Title>
+    </>
+  );
+};
+
+export default ChangePassword;
